+++ src/generators/Plugin.php
@@ -388,6 +388,7 @@ YAML;
     private function writePluginClass(): void
     {
         $file = new PhpFile();
+        $file->setStrictTypes($this->command->withStrictTypes);
 
         $namespace = $file->addNamespace($this->rootNamespace)
             ->addUse(Craft::class)
@@ -741,6 +742,7 @@ PHP
     private function writeSettingsModel(): void
     {
         $file = new PhpFile();
+        $file->setStrictTypes($this->command->withStrictTypes);
 
         $namespace = $file->addNamespace($this->settingsNamespace)
             ->addUse(Craft::class)